fix(C++): fix mismatched arg types to BMIN()/BMAX()/TCLAMP()
* lib/string.c:
    * str_nmove_ansi() len_rest (ssize_t -> size_t)
* maple/visio.c:
    * gety_ref() y_clipped (`unsigned int` -> int)
    * getx_ref() x_clipped (`unsigned int` -> int)
* so/pip.c:
    * pip_practice_class*() class_ (size_t -> int)
    * pip_practice_gradeup() (size_t -> int)
* so/pnote.c:
    * rebuild_pnote_ansi() `total` (size_t -> off_t)
    * do_pnote() `total` (size_t -> off_t)

+++ lib/string.c
@@ -517,7 +517,7 @@ GCC_PURE size_t str_nmove_ansi(const char *str, size_t idx, ssize_t diff, size_t
     {
         do
         {
-            const size_t len_rest = str_nlen(ptr, BMIN(diff, len));
+            const size_t len_rest = str_nlen(ptr, BMIN((size_t)diff, len));
             const char *const ptr_esc = (const char *)memchr(ptr, '\x1b', len_rest);
             const ssize_t step = (ptr_esc) ? ptr_esc - ptr : len_rest;
             ptr += step;
